renames request to friend-request and converts resource to typescript
