+++ controllers/friend-requests.ts
@@ -1,9 +1,10 @@
-const mongoose = require('mongoose');
-const Request = require('../models/request');
+import { Request, Response } from 'express';
+import mongoose from 'mongoose';
+import FriendRequest from '../models/friend-request';
 
 
 // create request
-async function createRequest (req, res) {
+export async function createRequest (req : Request, res : Response) {
     try {
         // verifies that both associated users have valid ids
         if (!mongoose.Types.ObjectId.isValid(req.body.from) || !mongoose.Types.ObjectId.isValid(req.body.to)) {
@@ -11,7 +12,7 @@ async function createRequest (req, res) {
         }
 
         // checks if request exists with count
-        const count = await Request.countDocuments({
+        const count = await FriendRequest.countDocuments({
             from: {$in: [req.body.from, req.body.to]},
             to: {$in: [req.body.from, req.body.to]}
         });
@@ -20,7 +21,7 @@ async function createRequest (req, res) {
         if (count) return res.status(409).json({ error: 'Duplicate request', reload: true });
         
         // creates a request if none exists
-        else await Request.create(req.body);
+        else await FriendRequest.create(req.body);
 
         res.redirect('/users/me');
     } catch (error) {
@@ -29,10 +30,10 @@ async function createRequest (req, res) {
 }
 
 // update request
-async function handleRequest (req, res) {
+export async function handleRequest (req : Request, res : Response) {
     try {
         // search by request id and userId to ensure request exists and the logged in user is authorized to make decision
-        const foundRequest = await Request.findOne({ _id: req.body.requestId, to: req.session.userId });
+        const foundRequest = await FriendRequest.findOne({ _id: req.body.requestId, to: req.session.userId });
         // else return error
         if (!foundRequest) return res.status(404).json({ error: 'Friend request not found', reload: true });
 
@@ -42,7 +43,7 @@ async function handleRequest (req, res) {
             await foundRequest.save();
         }
         // if user decline, delete request
-        else if (req.body.decision === 'Decline') await Request.findByIdAndDelete(foundRequest._id);
+        else if (req.body.decision === 'Decline') await FriendRequest.findByIdAndDelete(foundRequest._id);
         else return res.status(400).json({ error: 'Invalid decision for status update', reload: true });
         
         res.redirect('/users/me');
@@ -50,6 +51,3 @@ async function handleRequest (req, res) {
         res.status(500).send('An error occurred while updating the friend request');
     }
 }
-
-
-module.exports = {  createRequest, handleRequest }
